Add goal file decomposition
