+++ client/src/pages/profile/BasicInfoInput/styles.css
@@ -1,0 +1,16 @@
+.input-field {
+  position: relative;
+  align-items: center;
+  display: flex;
+  border: solid black 1px;
+  border-radius: 1rem;
+  width: 100%;
+  height: 13%;
+  margin-top: 5px;
+  padding: 3px;
+  font-size: 14px !important;
+}
+
+.input-field span {
+  flex-grow: 1;
+}
