+++ client/src/pages/profile/SocialsInput/styles.css
@@ -1,0 +1,8 @@
+.social-input {
+  padding: 0rem !important;
+  margin: 0rem;
+}
+
+.social-input .col {
+  padding: 5px !important;
+}
